commit in fabfile improve for issue #52
In the previous regex, it only concern the modified files, and in
commit e47e3a6, auto add untracked files is not a good choice, in
this commit I revert to only add tracked files;

And the issue mentioned the regex, I suggest to use `grep -E` instead of
`egrep`, for some system, `egrep` in not exists by default:

    add: git status --porcelain 2>/dev/null | grep -E "^ [MD]" | wc -l
    commit: git status --porcelain 2>/dev/null | grep -E "^[MADR]" | wc -l

But after severals attempts, I found use `git diff` to separate work tree and
index is simpler and better, and the `--exit-code` is easy to get the
diff status, there is a hack to not display the warning info by fabric.

In the end, the confirmed output is more preferred to git users, but not
common users, and I add a todo to format the output.